+++ crates/astria-composer/src/composer.rs
@@ -1,38 +1,16 @@
 use std::{
     io,
     net::SocketAddr,
-    time::Duration,
 };
 
-use astria_core::{
-    generated::composer::v1alpha1::{
-        composer_service_server::{
-            ComposerService,
-            ComposerServiceServer,
-        },
-        SubmitSequenceActionsRequest,
-    },
-    sequencer::v1alpha1::{
-        asset::default_native_asset_id,
-        transaction::action::SequenceAction,
-        RollupId,
-    },
-};
+use astria_core::generated::composer::v1alpha1::composer_service_server::ComposerServiceServer;
 use astria_eyre::eyre::{
     self,
     WrapErr as _,
 };
 use tokio::{
     net::TcpListener,
-    sync::mpsc::{
-        error::SendTimeoutError,
-        Sender,
-    },
     task::JoinError,
-};
-use tonic::{
-    Request,
-    Response,
 };
 use tracing::{
     error,
@@ -44,6 +22,7 @@ use crate::{
         self,
         ApiServer,
     },
+    composer_service::ExecutorHandle,
     searcher::Searcher,
     Config,
 };
@@ -64,11 +43,6 @@ pub struct Composer {
     grpc_collector_listener: TcpListener,
 }
 
-#[derive(Clone)]
-struct ExecutorHandle {
-    send_bundles: Sender<SequenceAction>,
-}
-
 impl Composer {
     /// Constructs a new Searcher service from config.
     ///
@@ -81,7 +55,7 @@ impl Composer {
             tokio::sync::mpsc::channel(256);
 
         let executor_handle = ExecutorHandle {
-            send_bundles: serialized_rollup_transactions_tx.clone(),
+            sequence_action_tx: serialized_rollup_transactions_tx.clone(),
         };
 
         let searcher = Searcher::from_config(
@@ -115,7 +89,8 @@ impl Composer {
     }
 
     /// Returns the socker address the grpc collector is served over
-    /// # Error: Returns an error if the listener is not bound
+    /// # Errors
+    /// Returns an error if the listener is not bound
     pub fn grpc_collector_local_addr(&self) -> io::Result<SocketAddr> {
         self.grpc_collector_listener.local_addr()
     }
@@ -147,9 +122,19 @@ impl Composer {
         });
 
         tokio::select! {
-            o = api_task => report_exit("api server", o),
-            o = searcher_task => report_exit("searcher", o),
-            o = grpc_server_handler => println!("grpc server ended: {:?}", o),
+                o = api_task => report_exit("api server", o),
+                o = searcher_task => report_exit("searcher", o),
+                o = grpc_server_handler => {
+                        match o {
+                            Ok(Ok(())) => info!(task = "grpc server", "task exited successfully"),
+                            Ok(Err(error)) => {
+                                error!(%error, task = "grpc_server", "task returned with error");
+                            }
+                            Err(error) => {
+                                error!(%error, task = "grpc server", "task failed to complete");
+                            }
+                        }
+                },
         }
     }
 }
@@ -163,47 +148,5 @@ fn report_exit(task_name: &str, outcome: Result<eyre::Result<()>, JoinError>) {
         Err(error) => {
             error!(%error, task = task_name, "task failed to complete");
         }
-    }
-}
-
-#[async_trait::async_trait]
-impl ComposerService for ExecutorHandle {
-    async fn submit_sequence_actions(
-        &self,
-        request: Request<SubmitSequenceActionsRequest>,
-    ) -> Result<Response<()>, tonic::Status> {
-        let submit_sequence_actions_request = request.into_inner();
-        if submit_sequence_actions_request.sequence_actions.is_empty() {
-            return Err(tonic::Status::invalid_argument(
-                "No sequence actions provided",
-            ));
-        }
-
-        // package the sequence actions into a SequenceAction and send it to the searcher
-        for sequence_action in submit_sequence_actions_request.sequence_actions {
-            let sequence_action = SequenceAction {
-                rollup_id: RollupId::from_unhashed_bytes(sequence_action.rollup_id),
-                data: sequence_action.tx_bytes,
-                fee_asset_id: default_native_asset_id(),
-            };
-
-            match self
-                .send_bundles
-                .send_timeout(sequence_action, Duration::from_millis(500))
-                .await
-            {
-                Ok(()) => {}
-                Err(SendTimeoutError::Timeout(_seq_action)) => {
-                    return Err(tonic::Status::deadline_exceeded(
-                        "timeout while sending txs to searcher",
-                    ));
-                }
-                Err(SendTimeoutError::Closed(_seq_action)) => {
-                    return Err(tonic::Status::unavailable("searcher is not available"));
-                }
-            }
-        }
-
-        Ok(Response::new(()))
     }
 }
